sep27

+++ Fase1/5.bFract'ol/MLX/fractol.h
@@ -6,7 +6,7 @@
 /*   By: fde-alen <fde-alen@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/04/25 13:20:58 by fde-alen          #+#    #+#             */
-/*   Updated: 2023/09/27 20:21:04 by fde-alen         ###   ########.fr       */
+/*   Updated: 2023/09/27 20:57:04 by fde-alen         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -154,8 +154,8 @@ typedef struct s_fractal
 	double		y;
 	double		x_shift;
 	double		y_shift;
-	double		mouse_x;
-	double		mouse_y;
+	int		mouse_x;
+	int		mouse_y;
 	double		arg_x;
 	double		arg_y;
 	double		radius;
